Path Nudge Fix
The nudge produced by the path correcting algorithm was adjusted to
prevent a forward boost created by the nudge if the user moved towards
the projected path. Only the perpendicular part of the vector is used to
steer the user onto the path. Also adjusted the name of the button to
better suit what each button is doing.