Implemented iteration 2 of map image to simulation
Created iteration 2 of the map image to simulation software which uses
OpenCV to generate polygons for the simulation software. In reality,
this wouldn't be necessary as the LiDAR data would be enough to steer
the user. However for the software, it is used to visualize the walls.
This iteration has not been integrated with the simulation and can be
ran as the file alone to test it. Maps folder has been added for any
map images used.